Fix the encoding of filenames and other parameters to exiftool.
So far, ``ExifTool.execute()`` expected all parameters to be Unicode
strings, and encoded them in UTF-8 before piping them to exiftool.  This
was wrong for two reasons: First, in Python 2.x filenames are raw strings
by convention, so no encoding is needed.  And second, the filenames should
be encoded using the system's filesystem encoding, which is not always
UTF-8 (unfortunately).

For a more consistent interface, a new method ``execute_json()`` has been
introduced that deals with the encodings.