commented out the fscodec() code that was first introduced in commit 843895ec6 back in 2012
I was unable to find the reference to where the code comes from the cpython source tree.  I would anticipate the old code from that commit was to deal with Python 2's str and unicode types that are no longer relevant.

tests still all pass.  If it ever is a problem, can add the code right back in since it was just commented out, not removed